+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { LocalizationProvider } from '@mui/x-date-pickers';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import { getCustomTheme, getMuiTheme, routesConfig } from './config';
 import { GlobalStyle } from './config';
@@ -12,6 +12,7 @@ import { CssBaseline, ThemeProvider as MuiThemeProvider } from '@mui/material';
 import { createTheme } from '@mui/material';
 
 import { useAppSelector } from './app/hooks';
+import { getAllBookings, getMyBookings } from './api/endpoints/ticket.api';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
@@ -25,8 +26,20 @@ function App() {
     const [loading] = useState(false);
     const mode = useAppSelector((state) => state.theme.currentTheme);
 
+    const test = async () => {
+        const res = await getMyBookings();
+        console.log(res);
+    };
+
+    useEffect(() => {
+        test().catch((e) => {
+            console.error(e);
+        });
+    }, []);
+
     return (
         <>
+            <ToastContainer theme={mode} />
             <LocalizationProvider dateAdapter={AdapterDateFns}>
                 <MuiThemeProvider theme={createTheme(getMuiTheme(mode))}>
                     <CustomThemeProvider theme={getCustomTheme(mode)}>
@@ -40,7 +53,6 @@ function App() {
                     </CustomThemeProvider>
                 </MuiThemeProvider>
             </LocalizationProvider>
-            <ToastContainer theme={mode} />
         </>
     );
 }