+++ src/api/converters/ticket.converter.ts
@@ -1,13 +1,18 @@
 import { IGender, ITicket } from '../../types';
-import { IBookingResponse } from '../types/ticket';
+import {
+    IBookingListingResponse,
+    IBookingResponse,
+    ITicketExternal,
+} from '../types/ticket';
+import { getTripFromTripExternal } from './trip.converter';
 
-export const getTicketFromBookingResponse = (
-    response: IBookingResponse
+const getTicketFromTicketExternals = (
+    ticketExternals: ITicketExternal[]
 ): ITicket => {
     const ticket: ITicket = {
-        pnrNumber: response[0].pnrNumber,
-        trip: response[0].trip,
-        seats: response.map((booking) => ({
+        pnrNumber: ticketExternals[0].pnrNumber,
+        trip: getTripFromTripExternal(ticketExternals[0].trip),
+        seats: ticketExternals.map((booking) => ({
             seatNumber: parseInt(booking.seatNumber),
             passenger: {
                 fullName: booking.passengerName,
@@ -17,4 +22,28 @@ export const getTicketFromBookingResponse = (
         })),
     };
     return ticket;
+};
+
+export const getTicketFromBookingResponse = (
+    response: IBookingResponse
+): ITicket => {
+    const ticket = getTicketFromTicketExternals(response);
+    return ticket;
+};
+
+export const getTicketsFromBookingListingResponse = (
+    response: IBookingListingResponse
+): ITicket[] => {
+    const ticketExternal = new Map<string, ITicketExternal[]>();
+    for (const booking of response.bookings) {
+        if (ticketExternal.has(booking.pnrNumber)) {
+            ticketExternal.get(booking.pnrNumber)?.push(booking);
+        } else {
+            ticketExternal.set(booking.pnrNumber, [booking]);
+        }
+    }
+    const tickets = Array.from(ticketExternal.values()).map((ticket) =>
+        getTicketFromTicketExternals(ticket)
+    );
+    return tickets;
 };