+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 -   [86cuhfyvp] Implemented functionality of Trip listing card using dummy data
 -   [86cugqgpt] Implemented filters and sort for action bar
 -   [86cueubmz] Implemented trip booking page and booking details form UI
+-   # [86cugexhf] Landing page illustration, action bar incorporated and routing to trip listing page.
 -   [86cukqvct] Implemented API layer
--   [86cugexhf] Landing page illustration, action bar incorporated and routing to trip listing page.
 -   [86cumak5j] Enhanced API layer and added converters for requests and responses
+-   [86cueujxt] Implemented bus search functionality for action bar and action bar filters, handling query params, corresponding reducer busSearchSlice and toggle animation
 -   [86cuewzcv] Implemented functionality of the booking page, added necessary toasts and confirmation dialog for booking.